sec1module: Split MultipleResultHelper into own file
