+++ zera-modules/sec1module/src/sec1modulemeasprogram.h
@@ -6,6 +6,7 @@
 #include "secinputinfo.h"
 #include "ecalcinfo.h"
 #include "clientactivecomponent.h"
+#include "multipleresulthelper.h"
 #include <stringvalidator.h>
 #include <basemeasprogram.h>
 #include <QDateTime>
@@ -207,91 +208,7 @@ private:
     quint32 m_nMeasurementNo;
     QTimer m_WaitMultiTimer;
     QDateTime m_WaitStartDateTime;
-    // TODO: Move MultipleResultHelper to a more common place
-    /**
-     * @brief Class MultipleResultHelper: Collect multple results / calculate statistics / out as JSON
-     */
-    class MultipleResultHelper {
-    public:
-        /**
-         * @brief clear: Clear resulr array / reset statistics
-         */
-        void clear();
-        /**
-         * @brief append: Add a reault to our array / update statistics
-         * @param fResult: Result value
-         * @param eRating: Result evaluation - see ECALCRESULT::enResultTypes
-         * @param fLowerLimit: Lower limit, the result was caclulated with
-         * @param fUpperLimit: Upper limit, the result was caclulated with
-         * @param fMaxError: Max error to set for e.g NaN
-         */
-        void append(const double fResult,
-                    const enum ECALCRESULT::enResultTypes eRating,
-                    const double fLowerLimit,
-                    const double fUpperLimit,
-                    const double fMaxError);
-        /**
-         * @brief getCountTotal
-         * @return Number of results in array
-         */
-        quint32 getCountTotal();
-        /**
-         * @brief getJSONResultArray: return JSON result array
-         * @note field names are as short as possible to reduce size
-         * @return JSON array as: { "V" : ResultValue,
-         *                          "R" : <Rating(Evaluation)>
-         *                          "T" : <Date/Time of result in short format 'yyyyMMddhhmmss'>
-         *                        }
-         */
-        const QJsonArray &getJSONResultArray();
-        /**
-         * @brief getJSONResultLimits: return JSON result limit array
-         * @note In most cases (constant limits during measurement) limit array
-         * has one entry only
-         * @return JSON array as: { "IDX": <start-index in results>
-         *                          "LL": <LowerLimit>,
-         *                          "UL": <UpperLimit>,
-         *                        }
-         */
-        const QJsonArray &getJSONLimitsArray();
-        /**
-         * @brief getJSONStatistics: return JSON object containing simple statistic values
-         * @return { "countPass: <count of passed results>,
-         *           "countFail": <count of passed results>,
-         *           "countUnfinish": <count of unfinished results>,
-         *           "mean": <mean (average) value,
-         *           "range": <maxError - minError>
-         *           "idxMin": <index of min error>
-         *           "idxMax": <index of max error>
-         *           "stddevN": <statndard deviantion / n in denominator,
-         *           "stddevN1": <statndard deviantion / (n-1) in denominator,
-         *          }
-         */
-        const QJsonObject getJSONStatistics();
-    private:
-        // mean / stddev
-        double m_fMeanValue = qQNaN();
-        double m_fStdDevn = qQNaN();
-        double m_fStdDevn1 = qQNaN();
-        // range
-        int m_iIdxMinValue = -1;
-        int m_iIdxMaxValue = -1;
-        double m_fMinError = qQNaN();
-        double m_fMaxError = qQNaN();
-        // results / counters
-        QJsonArray m_jsonResultArray;
-        int m_iCountPass = 0;
-        int m_iCountFail = 0;
-        int m_iCountUnfinish = 0;
-        // error estimation limits
-        QJsonArray m_jsonLimitArray;
-        double m_fLastLowerLimit = qQNaN();
-        double m_fLastUpperLimit = qQNaN();
-        /**
-         * @note To avoid one iteration of all results we keep accumulated sum
-         */
-        double m_fAccumulSum = 0.0;
-    };
+
     MultipleResultHelper m_multipleResultHelper;
 
     // methods
